change unsigned val to signed

+++ src/can_tools.cpp
@@ -20,7 +20,7 @@ CanDataTansformer::CanDataTansformer(int32_t start_bit, int32_t bit_length, doub
 {
 }
 
-bool CanDataTansformer::FromPhysicalData(const double src, uint8_t *dest, size_t len) const
+bool CanDataTansformer::FromPhysicalData(const double src, uint8_t *dest, int len) const
 {
     assert(dest && "fromPhysicalData:nullptr");
     if (src > _max_val || src < _min_val)
@@ -43,13 +43,13 @@ bool CanDataTansformer::FromPhysicalData(const double src, uint8_t *dest, size_t
 
     int pos_end = pos + step * _byte_covered_length;
 
-    if (pos_end < -1 || static_cast<size_t>(pos_end) > len)
+    if (pos_end < -1 || pos_end > len)
     {
         std::cerr << "warn:fromPhysicalData:Physical data is truncated,abort\n";
         return false;
     }
     pos_end = std::max(pos_end, -1);
-    pos_end = std::min(pos_end, static_cast<int>(len));
+    pos_end = std::min(pos_end, len);
 
     for (int i = 0; pos != pos_end; pos += step)
     {
@@ -58,7 +58,7 @@ bool CanDataTansformer::FromPhysicalData(const double src, uint8_t *dest, size_t
     return true;
 }
 
-bool CanDataTansformer::FromByteData(double &dest, const uint8_t *src, const size_t len) const
+bool CanDataTansformer::FromByteData(double &dest, const uint8_t *src, const int len) const
 {
     assert(src && "fromByteData:nullptr");
 
@@ -71,13 +71,13 @@ bool CanDataTansformer::FromByteData(double &dest, const uint8_t *src, const siz
 
     int pos_end = pos + step * _byte_covered_length;
 
-    if (pos_end < -1 || static_cast<size_t>(pos_end) > len)
+    if (pos_end < -1 || pos_end > len)
     {
         std::cerr << "warn:fromByteData:Physical data is truncated,abort\n";
         return false;
     }
     pos_end = std::max(pos_end, -1);
-    pos_end = std::min(pos_end, static_cast<int>(len));
+    pos_end = std::min(pos_end, len);
 
     for (int i = 0; pos != pos_end; pos += step)
     {
